Added two new functions...
Added functions TraitKnownByAllChars() and GetResearchedByAllColor() which are called from file "ResearchAssistant.lua".

New functions can be found at the end of the file.

At the moment this is just static. So there is no parameter in user settings yet for checking if trait is known by all characters. Also there is no parameter yet for custom color of traits known by all characters.